// src/types.ts
export interface Student {
  id: string;
  firstName: string;
  lastName: string;
  email: string;
  dateOfBirth: string;
  gender: string;
  phoneNumber: string;
  address: string;
  course: string;
  semester: string;
  enrollmentDate: string;
  bloodGroup: string;
  emergencyContact: string;
  hobbies: string[];
  profilePicture?: string;
}